requirements.txt
flask>=3.0.0
line-bot-sdk>=3.11.0
gspread>=6.0.0
google-auth>=2.28.0
gunicorn>=21.2.0
